implemented programable pixel shader stage in pipeline, added textureeffect, adapted scene
