+++ Engine/CubeSkinScene.h
@@ -4,11 +4,13 @@
 #include "Cube.h"
 #include "Mat3.h"
 #include "Pipeline.h"
+#include "TextureEffect.h"
 
 // scene demonstrating skinned cube
 class CubeSkinScene : public Scene
 {
 public:
+	typedef Pipeline<TextureEffect> Pipeline;
 	typedef Pipeline::Vertex Vertex;
 public:
 	CubeSkinScene( Graphics& gfx,const std::wstring& filename )
@@ -17,7 +19,7 @@ public:
 		pipeline( gfx ),
 		Scene( "Textured Cube skinned using texture: " + std::string( filename.begin(),filename.end() ) )
 	{
-		pipeline.BindTexture( filename );
+		pipeline.effect.ps.BindTexture( filename );
 	}
 	virtual void Update( Keyboard& kbd,Mouse& mouse,float dt ) override
 	{
